Add more lint items

+++ pyathena/pandas/util.py
@@ -60,7 +60,7 @@ def reset_index(df: "DataFrame", index_label: Optional[str] = None) -> None:
     try:
         df.reset_index(inplace=True)
     except ValueError as e:
-        raise ValueError(f"Duplicate name in index/columns: {e}")
+        raise ValueError("Duplicate name in index/columns") from e
 
 
 def as_pandas(cursor: "Cursor", coerce_float: bool = False) -> "DataFrame":
@@ -79,27 +79,25 @@ def to_sql_type_mappings(col: "Series") -> str:
     col_type = pd.api.types.infer_dtype(col, skipna=True)
     if col_type == "datetime64" or col_type == "datetime":
         return "TIMESTAMP"
-    elif col_type == "timedelta":
+    if col_type == "timedelta":
         return "INT"
-    elif col_type == "timedelta64":
+    if col_type == "timedelta64":
         return "BIGINT"
-    elif col_type == "floating":
+    if col_type == "floating":
         if col.dtype == "float32":
             return "FLOAT"
-        else:
-            return "DOUBLE"
-    elif col_type == "integer":
+        return "DOUBLE"
+    if col_type == "integer":
         if col.dtype == "int32":
             return "INT"
-        else:
-            return "BIGINT"
-    elif col_type == "boolean":
+        return "BIGINT"
+    if col_type == "boolean":
         return "BOOLEAN"
-    elif col_type == "date":
+    if col_type == "date":
         return "DATE"
-    elif col_type == "bytes":
+    if col_type == "bytes":
         return "BINARY"
-    elif col_type in ["complex", "time"]:
+    if col_type in ["complex", "time"]:
         raise ValueError(f"Data type `{col_type}` is not supported")
     return "STRING"
 
@@ -189,18 +187,17 @@ def to_sql(
     if if_exists == "fail":
         if table:
             raise OperationalError(f"Table `{schema}.{name}` already exists.")
-    elif if_exists == "replace":
-        if table:
-            cursor.execute(
-                textwrap.dedent(
-                    f"""
-                    DROP TABLE `{schema}`.`{name}`
-                    """
-                )
+    elif if_exists == "replace" and table:
+        cursor.execute(
+            textwrap.dedent(
+                f"""
+                DROP TABLE `{schema}`.`{name}`
+                """
             )
-            objects = bucket.objects.filter(Prefix=key_prefix)
-            if list(objects.limit(1)):
-                objects.delete()
+        )
+        objects = bucket.objects.filter(Prefix=key_prefix)
+        if list(objects.limit(1)):
+            objects.delete()
 
     if index:
         reset_index(df, index_label)
